[WIP][EasyAsync] Allow to have custom batch object
